refactor: refactoring UserDetails
UserDetails 클래스 도메인과 분리 -> 다형성 구현

related to #12
