+++ src/main/java/io/github/cokelee777/springsecurityjwtauth/domain/MemoryUser.java
@@ -1,11 +1,7 @@
 package io.github.cokelee777.springsecurityjwtauth.domain;
 
 import io.github.cokelee777.springsecurityjwtauth.enums.UserRole;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 
-import java.util.Collection;
-import java.util.List;
 import java.util.Objects;
 import java.util.UUID;
 
@@ -38,6 +34,18 @@ public class MemoryUser implements User {
         return identifier;
     }
 
+    public String getPassword() {
+        return password;
+    }
+
+    public UserRole getRole() {
+        return role;
+    }
+
+    public String getNickname() {
+        return nickname;
+    }
+
     private String createUUID(){
         return UUID.randomUUID().toString();
     }
@@ -57,40 +65,5 @@ public class MemoryUser implements User {
     @Override
     public int hashCode() {
         return Objects.hash(id, identifier, password, nickname, role);
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return List.of(new SimpleGrantedAuthority(this.role.toString()));
-    }
-
-    @Override
-    public String getPassword() {
-        return this.password;
-    }
-
-    @Override
-    public String getUsername() {
-        return this.identifier;
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
     }
 }
