+++ src/main/java/io/github/cokelee777/springsecurityjwtauth/domain/User.java
@@ -1,6 +1,4 @@
 package io.github.cokelee777.springsecurityjwtauth.domain;
 
-import org.springframework.security.core.userdetails.UserDetails;
-
-public interface User extends UserDetails {
+public interface User {
 }
