+++ src/main/java/io/github/cokelee777/springsecurityjwtauth/security/auth/JwtUserDetails.java
@@ -1,0 +1,6 @@
+package io.github.cokelee777.springsecurityjwtauth.security.auth;
+
+import io.github.cokelee777.springsecurityjwtauth.domain.User;
+
+public interface JwtUserDetails<T extends User> extends PrincipalUserDetails {
+}
