+++ src/main/java/io/github/cokelee777/springsecurityjwtauth/security/auth/PrincipalUserDetails.java
@@ -1,0 +1,6 @@
+package io.github.cokelee777.springsecurityjwtauth.security.auth;
+
+import org.springframework.security.core.userdetails.UserDetails;
+
+public interface PrincipalUserDetails extends UserDetails {
+}
