+++ src/main/java/io/github/cokelee777/springsecurityjwtauth/security/service/JwtMemoryUserDetailsService.java
@@ -2,6 +2,7 @@ package io.github.cokelee777.springsecurityjwtauth.security.service;
 
 import io.github.cokelee777.springsecurityjwtauth.domain.MemoryUser;
 import io.github.cokelee777.springsecurityjwtauth.repository.UserRepository;
+import io.github.cokelee777.springsecurityjwtauth.security.auth.JwtMemoryUserDetails;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 
@@ -15,7 +16,8 @@ public class JwtMemoryUserDetailsService implements JwtUserDetailsService<Memory
 
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        return userRepository.findByIdentifier(username).orElseThrow(
+        MemoryUser memoryUser = userRepository.findByIdentifier(username).orElseThrow(
                 () -> new UsernameNotFoundException("존재하지 않는 아이디 입니다"));
+        return new JwtMemoryUserDetails(memoryUser);
     }
 }
